Создан энумератор для размера (пока используется лишь для размера заказа, чуть позже добавится использование и для ячейки) Для единичной ячейки реализованы основные методы из абстрактного класса, но пока они не используются Теперь Cell абстрактный класс, а в коде используется single cell Для Order изменены использования конструктора в связи с добавлением энумератора размера Для ввода получателя в заказ добавлены поля и их обработка

+++ Postomat_App/Postomat_App/Postomat_App/CustomСontainer.cs
@@ -81,8 +81,8 @@ public class CustomСontainer<T> where T: Cell
             var orderString = cell.Content is null ? "" : cell.Content.GetOrderString();
             data.Add(new List<string>
             { 
-                cell.Identifier.ToString(), 
-                cell.Size.ToString(), 
+                cell.Identifier.ToString(),
+                ((int)cell.Size).ToString(), 
                 orderString
             });
         }
@@ -94,7 +94,7 @@ public class CustomСontainer<T> where T: Cell
     {
         foreach (var cell in Container)
         {
-            if (cell.Content is null && cell.Size >= order.Size)
+            if (cell.Content is null && cell.Size >= (int)order.Size)
             {
                 return cell.Identifier;
             }
